% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{permutation_example}
\alias{permutation_example}
\title{Permutation Example Result}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 7 columns.
}
\usage{
permutation_example
}
\description{
Results for a permutation example.
}
\details{
The code to generate this example permuation set can be found via
\code{system.file("example-scripts", "permutation_example.R", package = "pTITAN2")}
}
\seealso{
\code{vignette("pTITIAN2")}
}
\keyword{datasets}
