% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTITAN2.R
\name{ptitan}
\alias{ptitan}
\title{Permuted TITAN}
\usage{
ptitan(
  taxa,
  envs,
  sid,
  n = 6L,
  permutations = 100,
  titan_args = list(numPerm = 50, nBoot = 50)
)
}
\arguments{
\item{taxa}{a list of \code{data.frame}s with the taxa.}

\item{envs}{a list of \code{data.frame}s with the environmental gradients.}

\item{sid}{a character vector of length one with the name of the column
identifying the station id.}

\item{n}{the minimum number of occurrences, see \code{\link{occurrences}}.}

\item{permutations}{number of permutations to generate.}

\item{titan_args}{a list of named arguments to pass to \code{\link[TITAN2]{titan}}.}
}
\description{
A wrapper function to generate permustated results from
\code{\link[TITAN2]{titan}}.
}
\examples{

\dontrun{
eg <-
  ptitan(taxa = list(CD_06_Mall_wID, CN_06_Mall_wID, CN_06_Mall_wID, CN_06_Mall_wID),
         envs = list(C_IC_D_06_wID, C_IC_N_06_wID, C_IC_N_06_wID, C_IC_N_06_wID),
         sid  = "StationID",
         permutations = 20,
         titan_args = list(ncpus = 8, numPerm = 50, nBoot = 50, messaging = FALSE)
         )

eg

summary(eg)
}
}
\seealso{
\code{\link{permute}}, \code{\link{occurrences}},
\code{\link[TITAN2]{titan}}
}
