% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pTITAN2.R
\docType{package}
\name{pTITAN2}
\alias{pTITAN2}
\title{Permutations of Treatment Lables and TITAN2 Analysis}
\description{
Permutate treatment labels for taxa and environmental gradients to generate
an empirical distribution of change points.
}
